pyproject.toml
[tool.poetry]
name = "assurance"
version = "0.1.0"
description = "Insurance as a Service"
authors = ["Ankit Jain"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"
ruff = "^0.6.9"
django-stubs = "^5.1.0"
djangorestframework-stubs = "^3.15.1"
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
